tracker/pt: experimental non-white color removal
It's enabled only for automatic thresholding. With it on, it's possible
to keep tracking in normal light conditions without changing gain or
exposure beforehand.

It won't function on badly overexposed images, or with other bright
white colors in the frame. It should function on somewhat overexposed
images.

CPU usage is somewhat high, even taking advantage of all OpenCV SIMD
goodness as per the code.  We can revert the change if user reception is
bad.