tracker/pt: adjust auto threshold max pixel value
The confusion stems from the max slider value. It's 255, not 100 as is
the default. Max pixel count change to 20 was confused as we'll counting
area and not radius here. Taking into account that the value was divided
by 100 previously, it's bit less than the previous one but not as
confused as then.

Reported-by: @MathijsG
Issue: #355