pose-widget: avoid too small denominator in perpendicular alignment
With a very small denominator in barycentric coords formula we get a
dissociated octopus that took PCP and LSD in one sitting.

A dead Octopus is just a bunch of blue spheres.

IOW, in division with too small a denominator operand the results are
numerically unstable.

This is done in the constructor in a convoluted way as to avoid
branching during calculating for each pixel.

Issue: #356
Reported-by: @MathijsG