all: split "api" into "api" and "logic"
Here, the "logic" module has all the stuff for building one's own
graphical user interface.

The "api" module has stuff used by other trackers.

While at it, each of "api", "logic", and "compat" need their own export
headers. This is because of preprocessor symbol clashes.

This is all because a change in the "gui"-only dependency required a
relink of all the trackers, protocols, and flters. It takes too long
when building in the release configuration. With the split, only the
"gui" module gets rebuilt. Since it has close to no static dependencies,
it's fast enough.