cv/calibrator: limit samples at similar positions
Having yaw and pitch as a tuple, let N be the granularity. We're now only
allowing one sample per the granularity level.

Granularity -- "spacing_in_degrees" has a value of 3 degrees. For now the
values must be integral.

Since we're only allowing (yaw, pitch) tuples of given granularity, the
following get treated as distinct:

(0; 0), (0; 3), (0; 6), (1; 42), (3; 3)

The tuple value order can be swapped. There's nothing significant as for what's
pitch and what's yaw.

We drop the remainder between the yaw/pitch value so (0, 0) is index 0, (0; N)
is index 1, (0; 2N) index 3, etc.

This should prevent the calibration function from biasing itself when the user
keeps still during the procedure.