cmake, main: work toward introducing alternative UI
We're going to base opentrack derivatives on the same branch. Previously
merges were a living hell. Modularizing the UI code and having
continuously-built executables will do a lot.

First opentrack variant in progress is a TrackHat device for mouse and
scrolling control for people with spine and hand/arm disabilities.