tracker/wii: fix for mingw-w64
The module loses auto-pairing functionality, but it's
at least useful to see GCC warnings. I don't intend to
use GCC for actually running the module.

Fix some warnings emitted by GCC, such as

- memset(3) of non-POD struct
- depending on "new T[0]" to return a null pointer
- dependence on MSVC-specific pragmas for external
  libraries
- lack of header guards
- unintended usage of trigraph syntax

Disable some GCC warnings for the "wiiyourself"
library. It's not possible to patch it up, the code
quality is that awful.

Disable the builtin debug functionality since it
depends on non-standard __VA_OPT__ (due in C++20). It's
not possible to disable the warning emitted on the
__VA_ARGS__ usage in GCC.