video/widget: fix mutex starvation
Get rid of contention in `preview_size' and `set_image'.

After switching the Qt mutex to non-recursive, turns out the writer
thread preempts the UI thread to the point of freezing the entire thing.
Mutex fairness is an implementation detail and we must assume unfair
mutexes in the worst case.